Spring Boot: JPA oneToMany. [NEW] Section of the user quotes. [FEATURE] Quote editor.

+++ src/main/java/com/example/Quoter/repos/QuoteRepo.java
@@ -9,7 +9,7 @@ import com.example.Quoter.domain.Quote;
 //This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 //CRUD refers Create, Read, Update, Delete
 
-public interface QuoteRepo extends CrudRepository<Quote, Integer> {
+public interface QuoteRepo extends CrudRepository<Quote, Long> {
     
     List<Quote> findByTag(String tag);
